+++ src/canopy/knowledge_base/__init__.py
@@ -1,4 +1,3 @@
-from .knowledge_base import connect_to_pinecone
 from .knowledge_base import list_canopy_indexes
 from .knowledge_base import KnowledgeBase
 from .qdrant.qdrant_knowledge_base import QdrantKnowledgeBase