+++ src/canopy/knowledge_base/record_encoder/__init__.py
@@ -3,3 +3,5 @@ from .cohere import CohereEncoder
 from .dense import DenseRecordEncoder
 from .openai import OpenAIRecordEncoder
 from .anyscale import AnyscaleRecordEncoder
+from .azure_openai import AzureOpenAIRecordEncoder
+from .jina import JinaRecordEncoder