+++ src/canopy_cli/cli.py
@@ -18,8 +18,7 @@ import openai
 from openai import APIError as OpenAI_APIError
 from urllib.parse import urljoin
 
-from canopy.knowledge_base import KnowledgeBase
-from canopy.knowledge_base import connect_to_pinecone
+from canopy.knowledge_base import KnowledgeBase, list_canopy_indexes
 from canopy.knowledge_base.chunker import Chunker
 from canopy.chat_engine import ChatEngine
 from canopy.models.data_models import Document, UserMessage
@@ -41,13 +40,13 @@ load_dotenv()
 
 
 CONTEXT_SETTINGS = dict(help_option_names=['-h', '--help'])
-DEFAULT_SERVER_URL = f"http://localhost:8000/{API_VERSION}"
+DEFAULT_SERVER_URL = f"http://localhost:8000/{API_VERSION}/"
 spinner = Spinner()
 
 
 def check_server_health(url: str, timeout_seconds: int = 30):
     try:
-        res = requests.get(urljoin(url, "/health"), timeout=timeout_seconds)
+        res = requests.get(urljoin(url, "health"), timeout=timeout_seconds)
         res.raise_for_status()
         return res.ok
     except requests.exceptions.ConnectionError:
@@ -83,12 +82,11 @@ def wait_for_server(chat_server_url: str):
 
 def validate_pinecone_connection():
     try:
-        connect_to_pinecone()
+        list_canopy_indexes()
     except RuntimeError as e:
         msg = (
             f"{str(e)}\n"
-            "Credentials should be set by the PINECONE_API_KEY and PINECONE_ENVIRONMENT"
-            " environment variables.\n"
+            "Credentials should be set by the PINECONE_API_KEY environment variables.\n"
             "Please visit https://www.pinecone.io/docs/quickstart/ for more details."
         )
         raise CLIError(msg)
@@ -114,6 +112,13 @@ def _read_config_file(config_file: Optional[str]) -> Dict[str, Any]:
         raise CLIError(msg)
 
     return config
+
+
+def _load_create_index_params(config_file: Optional[str]) -> Dict[str, Any]:
+    config = _read_config_file(config_file)
+    if not config:
+        return {}
+    return config.get("create_index_params", {})
 
 
 def _load_kb_config(config_file: Optional[str]) -> Dict[str, Any]:
@@ -225,18 +230,21 @@ def health(url):
 @click.argument("index-name", nargs=1, envvar="INDEX_NAME", type=str, required=True)
 @click.option("--config", "-c", default=None, envvar="CANOPY_CONFIG_FILE",
               help="Path to a canopy config file. Can also be set by the "
-                   "`CANOPY_CONFIG_FILE` envrionment variable. Otherwise, the built-in"
-                   "defualt configuration will be used.")
+                   "`CANOPY_CONFIG_FILE` environment variable. Otherwise, the built-in"
+                   "default configuration will be used.")
 def new(index_name: str, config: Optional[str]):
     _initialize_tokenizer()
     kb_config = _load_kb_config(config)
     kb = KnowledgeBase.from_config(kb_config, index_name=index_name)
-    click.echo("Canopy is going to create a new index: ", nl=False)
-    click.echo(click.style(f"{kb.index_name}", fg="green"))
+    create_index_params = _load_create_index_params(config)
+    click.echo("Canopy is going to create a new index named ", nl=False)
+    click.echo(click.style(f"{kb.index_name}", fg="green"), nl=False)
+    click.echo(" with the following initialization parameters:")
+    click.echo(click.style(f"{yaml.dump(create_index_params)}", fg="cyan"))
     click.confirm(click.style("Do you want to continue?", fg="red"), abort=True)
     with spinner:
         try:
-            kb.create_canopy_index()
+            kb.create_canopy_index(**create_index_params)
         # TODO: kb should throw a specific exception for failure
         except Exception as e:
             already_exists_str = f"Index {kb.index_name} already exists"
@@ -280,6 +288,9 @@ def _batch_documents_by_chunks(chunker: Chunker,
 
         Load all the documents from a data file or a directory containing multiple data
         files. The allowed formats are .jsonl, .parquet, .csv, and .txt.
+
+        If you would like partition your data into namespaces, you can use the namespace parameter.
+        For more information see: https://docs.pinecone.io/docs/namespaces
         """  # noqa: E501
     )
 )
@@ -297,12 +308,17 @@ def _batch_documents_by_chunks(chunker: Chunker,
                    "long as less than 10% of the documents have failed to be uploaded.")
 @click.option("--config", "-c", default=None, envvar="CANOPY_CONFIG_FILE",
               help="Path to a canopy config file. Can also be set by the "
-                   "`CANOPY_CONFIG_FILE` envrionment variable. Otherwise, the built-in"
-                   "defualt configuration will be used.")
+                   "`CANOPY_CONFIG_FILE` environment variable. Otherwise, the built-in "
+                   "default configuration will be used.")
+@click.option("--namespace", "-n", default="", envvar="INDEX_NAMESPACE",
+              help="The namespace of the index. Can also be set by the "
+                   "`INDEX_NAMESPACE` environment variable. If not set, the default "
+                   "namespace will be used.")
 def upsert(index_name: str,
            data_path: str,
            allow_failures: bool,
-           config: Optional[str]):
+           config: Optional[str],
+           namespace: str):
     if index_name is None:
         msg = (
             "No index name provided. Please set --index-name or INDEX_NAME environment "
@@ -330,7 +346,10 @@ def upsert(index_name: str,
     click.echo("Canopy is going to upsert data from ", nl=False)
     click.echo(click.style(f'{data_path}', fg='yellow'), nl=False)
     click.echo(" to index: ")
-    click.echo(click.style(f'{kb.index_name} \n', fg='green'))
+    click.echo(click.style(f'{kb.index_name}', fg='green'), nl=False)
+    click.echo(" using namespace: ", nl=False)
+    click.echo(click.style(f'{namespace or "default"} \n', fg='cyan'))
+
     with spinner:
         try:
             data = load_from_path(data_path)
@@ -366,7 +385,7 @@ def upsert(index_name: str,
     for batch in _batch_documents_by_chunks(kb._chunker, data,
                                             batch_size=kb._encoder.batch_size):
         try:
-            kb.upsert(batch)
+            kb.upsert(batch, namespace=namespace)
         except Exception as e:
             if allow_failures and len(failed_docs) < len(data) // 10:
                 failed_docs.extend([_.id for _ in batch])
@@ -402,7 +421,7 @@ def _chat(
     openai_api_key=None,
     api_base=None,
     stream=True,
-    print_debug_info=False,
+    print_debug_info=False
 ):
     if openai_api_key is None:
         openai_api_key = os.environ.get("OPENAI_API_KEY")
@@ -413,9 +432,11 @@ def _chat(
             "Please set the OPENAI_API_KEY environment "
             "variable."
         )
+
+    client = openai.OpenAI(base_url=api_base, api_key=openai_api_key)
+
     output = ""
     history += [{"role": "user", "content": message}]
-    client = openai.OpenAI(base_url=api_base, api_key=openai_api_key)
 
     start = time.time()
     try:
@@ -431,6 +452,8 @@ def _chat(
     click.echo(click.style(f"\n {speaker}:\n", fg=speaker_color))
     if stream:
         for chunk in openai_response:
+            if not chunk.choices:
+                continue
             openai_response_id = chunk.id
             internal_model = chunk.model
             text = chunk.choices[0].delta.content or ""
@@ -478,6 +501,7 @@ def _chat(
         """
 
     )
+
 )
 @click.option("--stream/--no-stream", default=True,
               help="Stream the response from the RAG chatbot word by word.")
@@ -581,6 +605,12 @@ def chat(chat_server_url, rag, debug, stream):
 
         If you would like to try out the chatbot, run `canopy chat` in a separate
         terminal window.
+
+        You can also use the OpenAPI client to access the Canopy server just
+        by changing the API endpoint.
+
+        For more information see:
+        https://github.com/pinecone-io/canopy?tab=readme-ov-file#migrating-an-existing-openai-application-to-canopy
         """
     )
 )
@@ -675,7 +705,7 @@ def stop(url):
                     raise CLIError(msg)
 
     try:
-        res = requests.get(urljoin(url, "/shutdown"))
+        res = requests.get(urljoin(url, "shutdown"))
         res.raise_for_status()
         return res.ok
     except requests.exceptions.ConnectionError:
@@ -693,8 +723,8 @@ def stop(url):
         """
     )
 )
-@click.option("--url", default="http://localhost:8000",
-              help="Canopy's server url. Defaults to http://localhost:8000")
+@click.option("--url", default=DEFAULT_SERVER_URL,
+              help=f"Canopy's server url. Defaults to {DEFAULT_SERVER_URL}")
 def api_docs(url):
     import webbrowser
 
@@ -720,7 +750,7 @@ def api_docs(url):
             print(HTML_TEMPLATE % json.dumps(app.openapi()), file=fd)
         webbrowser.open('file://' + os.path.realpath(filename))
     else:
-        webbrowser.open(urljoin(url, "redoc"))
+        webbrowser.open(urljoin(url, "/redoc"))
 
 
 if __name__ == "__main__":